fix(composable): match the Li float ABI in lig/studio externs and the sim fps sample
Two runtime contract violations surfaced by the composable probes:
li_rt_lig_kernel_last_validity_ratio and the three studio/lig `-> float`
externs returned C `float` while Li `float` is 64-bit, so callers read f32
bits as an f64 denormal (1.0f came back as 5.26e-315, failing the validity
gate). They now return double. Separately, the 60 Hz frame-time sample
(16.667 ms) made render_bench_fps_counter_simulate land at ~59.9988 fps, just
under its own meets_target == 1 contract, so the viewport FPS probe exited 5;
the sample now stays inside the exact 1000/60 ms budget.

+++ runtime/li_rt.c
@@ -1510,7 +1510,9 @@ int32_t li_rt_studio_timeline_tick_frame(void) {
   return 1;
 }
 
-float li_rt_studio_timeline_playhead_pct(void) { return g_studio_timeline_playhead_pct; }
+/* Li `float` is 64-bit (float64): externs declared `-> float` must return
+ * double, not C float (a f32 return read as f64 yields a denormal). */
+double li_rt_studio_timeline_playhead_pct(void) { return (double)g_studio_timeline_playhead_pct; }
 
 int32_t li_rt_studio_timeline_reset_mock(void) {
   g_studio_timeline_playing = 0;
@@ -1635,12 +1637,12 @@ int32_t li_rt_lig_host_present_active(void) {
   return g_lig_host_present_active;
 }
 
-float li_rt_lig_host_present_dt_ms(void) {
+double li_rt_lig_host_present_dt_ms(void) {
   li_rt_lig_refresh_host_active();
   if (g_lig_host_present_active) {
-    return g_lig_present_dt_ms;
+    return (double)g_lig_present_dt_ms;
   }
-  return 16.667f;
+  return 16.667;
 }
 
 int32_t li_rt_lig_host_native_pixels(void) {
@@ -1685,11 +1687,11 @@ int32_t li_rt_lig_wgpu_present_frame(int32_t swapchain_ok) {
 int32_t li_rt_studio_shell_input_pointer_down(void) {
   return g_studio_shell_pointer_down;
 }
-float li_rt_studio_shell_input_pointer_x(void) {
-  return g_studio_shell_pointer_x;
+double li_rt_studio_shell_input_pointer_x(void) {
+  return (double)g_studio_shell_pointer_x;
 }
-float li_rt_studio_shell_input_pointer_y(void) {
-  return g_studio_shell_pointer_y;
+double li_rt_studio_shell_input_pointer_y(void) {
+  return (double)g_studio_shell_pointer_y;
 }
 int32_t li_rt_studio_shell_input_key_escape(void) {
   return g_studio_shell_key_escape;
