+++ runtime/li_rt_lig.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 int32_t li_rt_lig_kernel_run(int32_t kernel_id, int32_t backend_id);
-float li_rt_lig_kernel_last_validity_ratio(void);
+double li_rt_lig_kernel_last_validity_ratio(void);
 
 #endif
